test: add edgeless test action module
