+++ tests/utils/actions/edgeless.ts
@@ -1,0 +1,26 @@
+/* eslint-disable @typescript-eslint/no-restricted-imports */
+import '../declare-test-window.js';
+
+import { expect, Page } from '@playwright/test';
+
+import type { FrameBlockModel } from '../../../packages/blocks/src/index.js';
+
+export async function getFrameSize(
+  page: Page,
+  ids: { pageId: string; frameId: string; paragraphId: string }
+) {
+  const result: string | null = await page.evaluate(
+    ([id]) => {
+      const page = window.workspace.getPage('page0');
+      const block = page?.getBlockById(id.frameId);
+      if (block?.flavour === 'affine:frame') {
+        return (block as FrameBlockModel).xywh;
+      } else {
+        return null;
+      }
+    },
+    [ids] as const
+  );
+  expect(result).not.toBeNull();
+  return result as string;
+}
